src/main/java/org/hyp/springboot/rest/jpa/service/CityService.java
package org.hyp.springboot.rest.jpa.service;

import org.hyp.springboot.rest.jpa.domain.City;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.util.Assert;

@Service
public class CityService {
    @Autowired
    CityRepository cityRepository;

    public City getCity(String name, String country) {
        Assert.notNull(name, "Name must not be null");
        Assert.notNull(country, "Country must not be null");
        return this.cityRepository.findByNameAndCountryAllIgnoringCase(name, country);
    }
}
